+++ rest-collections/Kentobot API/Save Song Play.bru
@@ -13,11 +13,9 @@ post {
 body:json {
   {
     "title": "Song Title 2",
-    "requestedBy": "jsnphil",
-    "lenght": 180,
-    "played": "2024-11-23T01:09:05.456Z",
+    "requestedBy": "user3",
+    "length": 180,
+    "played": "2024-11-25T01:09:05.456Z",
     "youtubeId": "youtube-id"
   }
-  
-  
 }
